Additional logging, introduce gc_color_purple
Regarding the purple color, it looks like when marking we will need to check for that, too. The problem is that the thread data takes into account overlaps where more than one clear color is in play, but the collector does not at this time. It can use the purple color to do that, and save objects that are not quite garbage yet.